GEODE-6786: Add a few more tests to validate region delete status codes (#3685)
